move zoomtodisplay into LGV model, start on matching refname
